requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12
python-dotenv==1.0.1
